// frontend/src/api-generated/models/phieuCapPhatTaiSanResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietCapPhatGeneralResponse } from './chiTietCapPhatGeneralResponse';

export interface PhieuCapPhatTaiSanResponse {
  id?: number;
  idDonVi?: number;
  maPhiepCapPhat?: string;
  idNguoiNhan?: number;
  tenNguoiNhan?: string;
  idPhongBanNhan?: number;
  tenPhongBanNhan?: string;
  idNguoiLap?: number;
  tenNguoiLap?: string;
  idNguoiPheDuyet?: number;
  tenNguoiPheDuyet?: string;
  thoiGianBanGiao?: string;
  trangThai?: string;
  mucDichSuDung?: string;
  danhSachTaiSan?: ChiTietCapPhatGeneralResponse[];
  thoiGianTao?: string;
  thoiGianCapNhat?: string;
}