// frontend/src/api-generated/models/phieuCapPhatTaiSanRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietCapPhatLinhKienRequest } from './chiTietCapPhatLinhKienRequest';
import type { ChiTietCapPhatPhanCungRequest } from './chiTietCapPhatPhanCungRequest';
import type { ChiTietCapPhatPhanMemRequest } from './chiTietCapPhatPhanMemRequest';

export interface PhieuCapPhatTaiSanRequest {
  idNguoiNhan: number;
  idPhongBanNhan: number;
  mucDichSuDung?: string;
  danhSachPhanCung?: ChiTietCapPhatPhanCungRequest[];
  danhSachPhanMem?: ChiTietCapPhatPhanMemRequest[];
  danhSachLinhKien?: ChiTietCapPhatLinhKienRequest[];
}